feat: 카테고리 리스트 조회/카테고리 등록/카테고리 수정/카테고리 삭제
카테고리 수정/삭제는 오류가 일어날 수도 있음 후에 더 체크할 예정

resolve: #32,#33,#34,#35

+++ src/main/java/com/kiosk/kioskback/service/implementation/CategoryServiceImplements.java
@@ -17,6 +17,7 @@ import com.kiosk.kioskback.entity.CategoryEntity;
 import com.kiosk.kioskback.entity.StoreEntity;
 import com.kiosk.kioskback.entity.UserEntity;
 import com.kiosk.kioskback.repository.CategoryRepository;
+import com.kiosk.kioskback.repository.MenuRepository;
 import com.kiosk.kioskback.repository.StoreRepository;
 import com.kiosk.kioskback.repository.UserRepository;
 import com.kiosk.kioskback.service.CategoryService;
@@ -27,6 +28,7 @@ public class CategoryServiceImplements implements CategoryService {
     @Autowired CategoryRepository categoryRepository;
     @Autowired UserRepository userRepository;
     @Autowired StoreRepository storeRepository;
+    @Autowired MenuRepository menuRepository;
 
     public ResponseDto<List<GetCategoryResponseDto>> getList(int storeId) {
         List<GetCategoryResponseDto> data = null;
@@ -74,15 +76,14 @@ public class CategoryServiceImplements implements CategoryService {
 
             CategoryEntity categoryEntity = categoryRepository.findByCategoryId(categoryId);
             if (categoryEntity == null) return ResponseDto.setFailed("Not Exist Category");
-
-            List<StoreEntity> storeEntity = storeRepository.findByUserId(userId);
-            boolean isPermission = userId.equals(storeEntity);
-            if (!isPermission) return ResponseDto.setFailed("Do Not Permmission!");       
+   
+            UserEntity userEntity = userRepository.findByUserId(userId);
+            if (!userEntity.isAdmin()) return ResponseDto.setFailed("Do Not Admin!");         
 
             categoryEntity.patch(dto);
             categoryRepository.save(categoryEntity);
 
-            data = new PatchCategoryResponseDto(categoryId, userId, categoryId);         
+            data = new PatchCategoryResponseDto(categoryId, userId, categoryId);
 
         } catch (Exception exception) {
             exception.printStackTrace();
@@ -100,8 +101,10 @@ public class CategoryServiceImplements implements CategoryService {
             CategoryEntity categoryEntity = categoryRepository.findByCategoryId(categoryId);
             if (categoryEntity == null) return ResponseDto.setFailed("Not Exist CategoryId");
 
-            // boolean isEqualWriteUser = userId.equals(categoryEntity.getWriteUser());
-            // if (!isEqualWriteUser) return ResponseDto.setFailed("Do Not Permission");
+            UserEntity userEntity = userRepository.findByUserId(userId);
+            if (!userEntity.isAdmin()) return ResponseDto.setFailed("Do Not Admin!");
+
+            menuRepository.deleteByCategoryId(categoryId);
 
             categoryRepository.delete(categoryEntity);
